pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "check404"
version = "1.0.0"
description = "Simple tool for checking code correctness"
readme = "README.rst"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [{name = "Rodrigo Nicola", email = "rnicola@maua.br"}]
keywords = ["test", "didatic"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Education",
  "License :: Freely Distributable",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development :: Testing",
  "Typing :: Typed"
]
dependencies = ["patool>=1.10", "pyyaml>=6.0"]
#dynamic = ["version"]

[project.scripts]
check404 = "check404.checker:check"

[project.urls]
Source = "https://github.com/ECM404/check404"

[tools.setuptools.packages]
find = {}
